refactor: 🎨 moveditsection moved to common components
